requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.12
google-generativeai==0.8.4
grailed_api==0.1.5
mercari==2.2.1
PyYAML==6.0.2
python-dateutil==2.9.0
requests==2.32.3
resend==2.5.1
responses==0.25.3
pytest==8.3.4
pytest-cov==6.0.0
ruff==0.9.6
